WIP: add traces.py
